# backend/api/openapi.yaml
...
openapi: 3.1.0
...
info:
  title: Foraged goods marketplace API
  version: 0.1.0
  # "|" keeps the newlines - the description renders as Markdown in the UI.
  description: |
    A marketplace for foraged goods — berries, mushrooms, herbs. Sellers post
    listings, buyers browse, message them and place orders. Payments are
    simulated; no money changes hands.

    ## Authentication

    Two tokens, with opposite properties.

    The **access token** is a signed JWT sent on every request:

        Authorization: Bearer <access_token>

    It is verified without touching the database, which also means it cannot be
    revoked - so it is short-lived (15 minutes by default, `ACCESS_TOKEN_TTL`).

    The **refresh token** is an opaque value in an HttpOnly `refresh_token`
    cookie, set by signup and login and valid for 7 days. `POST /auth/refresh`
    exchanges it for a new access token and rotates it. Only its hash is
    stored, so a database leak cannot resume a session.

    When an access token expires, the 401 carries
    `WWW-Authenticate: Bearer error="invalid_token"` - the signal to refresh
    rather than to show a login screen. `GET /auth/me` returns the identity
    behind a token that is still valid, for restoring a session on page load.

    Public endpoints (browsing listings, search, images) work without one.

    ## Ids

    Every id is a **uuid v7**: the leading 48 bits are a millisecond timestamp,
...
    something to parse.

    ## Errors

...

...

...
servers:
...

...
security:
  - bearerAuth: []
  - apiKeyAuth: []

# "components" is the reusable-definitions section. Nothing in here appears in
# the docs on its own - it only shows up where something $refs it.
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        A key from `POST /me/api-keys`, for machine clients. It acts as its
        owner - same permissions, same ownership checks - and is rate limited
        per key rather than per IP.

        Keys cannot manage keys: creating, listing and revoking need a session,
        so a leaked key cannot mint a replacement and survive its own
        revocation.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # documentation only; no tool verifies the format
      description: |
        Access token issued by signup or login. Expired and missing tokens are
        both answered with 401.
  schemas:
...
    Error:
      type: object
      required: [error] # "details" is omitempty in Go, so it is not required
      properties:
        error:
          type: string
          description: |
            Safe to show a user on 4xx, and written for display: sentence case,
            no trailing full stop. Render it as-is rather than transforming it.
            On 5xx this is always "Something went wrong" — the cause is in the
            server log, tagged with the `X-Request-ID` on the response.
          examples: ["Listing not found"]
        details:
          type: object
          # "any key, string value" - the YAML spelling of map[string]string.
          additionalProperties:
            type: string
          description: Reserved for field-level validation errors; absent today.
    ListingImage:
      type: object
      required: [id, url, position]
      properties:
        id:
          type: string
          format: uuid
          examples: ["01a02305-b81c-7dcb-86a0-7f75e33e0af3"]
        url:
          type: string
          description: Path served by the API, e.g. for use in an `img` tag.
          examples: ["/uploads/0a7a0bc6-99f9-42f8-9ba0-c908a9cb3a93.png"]
        position:
          type: integer
          format: int32
          description: Display order, ascending. Unique per listing.
          examples: [0]
    Listing:
      type: object
      # Every field without omitempty in ListingResponse is always present.
      required:
        [
          id,
          seller_id,
          title,
          description,
          category,
          price,
          quantity,
          unit,
          created_at,
          updated_at,
          images,
        ]
      properties:
        id:
          type: string
          format: uuid
          examples: ["01a02305-b81c-7dcb-86a0-7f75e33e0af3"]
        seller_id:
          type: string
          format: uuid # Go uuid.UUID marshals to a string
        title:
          type: string
          maxLength: 100
          examples: ["Golden Chanterelles"]
        description:
          type: string
          description: Empty string when the seller left it blank.
        category:
          type: string
          examples: ["mushrooms"]
        price:
          type: number
          format: double
          description: Price per unit. Stored as NUMERIC(10,2).
          examples: [18.5]
        quantity:
          type: integer
          format: int32
          description: Remaining quantity; 0 means sold out.
          examples: [4]
        unit:
          type: string
          examples: ["kg"]
        created_at:
          type: string
          format: date-time # Go time.Time marshals to RFC 3339
        updated_at:
          type: string
          format: date-time
        images:
          type: array
          description: Always present; empty when the listing has no photos.
          items:
            # $ref = "reuse that definition here". This is why schemas exist.
            $ref: "#/components/schemas/ListingImage"
...
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json: # the media type the body is sent as
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: |
        No token, or one that is expired or invalid. The body is the same in
        every case; `WWW-Authenticate` says which, because a client's response
        differs: refresh, or log in.
      headers:
        WWW-Authenticate:
          description: |
            `Bearer error="invalid_token", error_description="The access token expired"`
            when the token was valid but has aged out - retry through
            `POST /auth/refresh`. A bare `Bearer` for a missing, malformed or
            forged token - the client should log in. A forged token reads as
            absent on purpose.
          schema:
            type: string
          examples:
            expired:
              value: Bearer error="invalid_token", error_description="The access token expired"
            absent:
              value: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: |
        No such resource — also returned when a resource exists but the caller
        is not a participant, so its existence is not revealed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: |
        Too many requests on this API key. Only key-authenticated requests are
        limited; browser sessions are not.
      headers:
        Retry-After:
          description: Seconds until a request will be accepted again.
          schema:
            type: integer
            examples: [2]
        X-RateLimit-Limit:
          description: |
            The bucket's capacity. Sent on every key-authenticated response,
            not only on a refusal.
          schema:
            type: integer
            examples: [60]
        X-RateLimit-Remaining:
          description: Requests left in the current bucket.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: Something failed on our side; the cause is in the log.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# Each key is a URL relative to the server above, so "/listings" is really
# http://localhost:8080/api/v1/listings.
paths:
...
  /listings/{id}/save:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [wishlist]
      summary: Save a listing
      operationId: saveListing
      description: |
        Idempotent — saving a listing you already saved is still a 204, not a
        409 (the insert is `ON CONFLICT DO NOTHING`).
      responses:
        "204":
          description: Saved. No body.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [wishlist]
      summary: Remove a listing from the wishlist
      operationId: unsaveListing
      description: |
        NOT idempotent, unlike save: removing something you had not saved is a
        404 ("listing was not saved").
      responses:
        "204":
          description: Removed. No body.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /me/saved:
    get:
      tags: [wishlist, me]
      summary: List your saved listings
      operationId: getSavedListings
      description: Unpaginated. Photos are attached in one batch query.
      responses:
        "200":
          description: Your wishlist.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"

  # ---------------------------------------------------------------------------
  # orders (simulated - no real money moves)
  # ---------------------------------------------------------------------------
...